Refactor the masks functions on a new unit.
This is a significant change, as the new mask functions are all refactored
to use the new primitives, and that requires changing the definition
of the previously-global variables.

On the other hand, this completes the refactorign of imageprocess.c
and removes the need for any globals in the imageprocess side.