Calculate blackfilter exclusions as Rectangles to begin with.
This removes the remaining maskToRectangle() usage, so that this
convenience function can go away.

The logic in place for auto-selecting the exclusion zone is actually different,
in so far as it provides a more correct selection based on size, rather
than being off-by-one due to the lack of `-1` in the calculation.